+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Brian2694\Toastr\Facades\Toastr;
 use Illuminate\Http\Request;
@@ -98,6 +99,16 @@ class SubCategoryController extends Controller
   public function destroy(string $id)
   {
     $subCategory = SubCategory::findOrFail($id);
+
+    $childCategoryCount = ChildCategory::where("sub_category_id", $subCategory->id)->count();
+
+    if ($childCategoryCount > 0) {
+      return response([
+        "message" => "Danh mục cấp 2 này đang có ít nhất 1 danh mục cấp 3, vui lòng xoá các danh mục cấp 3 trước",
+        "status" => "error"
+      ]);
+    }
+
     $subCategory->delete();
 
     return response([
